fix(coeus-ops): Gate ConvTranspose3d fallback
Move 3-D transposed convolution behind a dedicated capability trait so CPU defaults dispatch through Leto and future accelerator providers can own native Hephaestus kernels without a host-copy fallback.
