+++ crates/coeus-ops/src/backend_ops/cpu_impl/impls/mod.rs
@@ -1,5 +1,6 @@
 mod attention;
 mod conv;
+mod conv_transpose3d;
 mod elementwise;
 mod matmul;
 mod optim;
