+++ crates/coeus-ops/src/backend_ops/traits/mod.rs
@@ -11,7 +11,8 @@
 //! - [`ElementwiseOps`] — binary and unary element-wise ops
 //! - [`MatmulOps`] — matmul, batched matmul, accumulate variants
 //! - [`ReductionOps`] — reduce, argmax/argmin, topk, cumulative sum/product scans
-//! - [`ConvOps`] — 1D/2D/3D conv forward+backward, transposed conv defaults
+//! - [`ConvOps`] — regular convolution and 1D/2D transposed convolution
+//! - [`ConvTranspose3dOps`] — backend-owned 3-D transposed convolution
 //! - [`PoolOps`] — max/avg pool 1D/2D/3D forward+backward
 //! - [`AttentionOps`] — scaled dot-product attention forward+backward
 //! - [`OptimizerOps`] — fused SGD/Adam/RMSProp/AdamW/AdaGrad steps
@@ -21,6 +22,7 @@
 
 pub mod attention;
 pub mod conv;
+pub mod conv_transpose3d;
 pub mod elementwise;
 pub mod matmul;
 pub mod optimizer;
@@ -30,6 +32,7 @@ pub mod unfold_fold;
 
 pub use attention::AttentionOps;
 pub use conv::ConvOps;
+pub use conv_transpose3d::ConvTranspose3dOps;
 pub use elementwise::ElementwiseOps;
 pub use matmul::MatmulOps;
 pub use optimizer::OptimizerOps;
